server/requirements.txt
torchvision==0.6.0
tensorflow==2.1.0
nltk==3.5
keras==2.3.1
Flask==1.1.2
